Move streamingpromql engine opts into its own package
