+++ pkg/streamingpromql/opts/config.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package streamingpromql
+package opts
 
 import (
 	"flag"
